Revert "Rename lib/translationunit.js to lib/translation-unit.js"
This reverts commit e2c8421980bcdca425316904be0e2d2cab7814df.
